refactor: merged CreateEvent and EditEvent components into a single EventEditor component
