--- src/components/EventEditor/CreateEvent/index.js
@@ -1,3 +1,0 @@
-import CreateEvent from './CreateEvent';
-
-export default CreateEvent;
