--- src/components/EventEditor/EditEvent/edit-event.css
@@ -1,5 +1,0 @@
-.option-wrapper {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
