--- src/components/EventEditor/EditEvent/index.js
@@ -1,3 +1,0 @@
-import EditEvent from './EditEvent';
-
-export default EditEvent;
